+++ quantrl/plotters.py
@@ -6,7 +6,7 @@
 __name__    = 'quantrl.plotters'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2023-12-08"
-__updated__ = "2024-05-30"
+__updated__ = "2024-07-22"
 
 # dependencies
 from io import BytesIO
@@ -23,8 +23,17 @@ os.environ['KMP_DUPLICATE_LIB_OK'] = 'TRUE'
 plt.rcParams['font.family'] = "Times New Roman"
 plt.rcParams['font.size'] = 16
 plt.rcParams['mathtext.fontset'] = 'cm'
-plt.rcParams['figure.subplot.hspace'] = 0.2
-plt.rcParams['figure.subplot.wspace'] = 0.25
+plt.rcParams['axes.grid'] = True
+plt.rcParams['grid.alpha'] = 0.5
+plt.rcParams['grid.linewidth'] = 0.5
+plt.rcParams['xtick.top'] = True
+plt.rcParams['xtick.minor.visible'] = True
+plt.rcParams['ytick.right'] = True
+plt.rcParams['ytick.minor.visible'] = True
+plt.rcParams['axes.titlesize'] = 16
+plt.rcParams['axes.labelsize'] = 16
+plt.rcParams['xtick.labelsize'] = 14
+plt.rcParams['ytick.labelsize'] = 14
 
 class TrajectoryPlotter(object):
     """Plotter for trajectories.
@@ -34,7 +43,7 @@ class TrajectoryPlotter(object):
     Parameters
     ----------
     axes_args: list
-        Lists of axis properties. The first element of each entry is the ``x_label``, the second is ``y_label``, the third is ``[y_limit_min, y_limit_max]`` and the fourth is ``y_scale``. Default is ``[]``.
+        Lists of axis properties. The first element of each entry is the ``x_label``, the second is ``y_label``, the third is ``[y_limit_min, y_limit_max]`` and the fourth is ``y_scale``.
     axes_lines_max: int, default=10
         Maximum number of lines to display in each plot. Higher numbers slow down the run. Default is ``10``.
     axes_cols: int, default=2
@@ -207,6 +216,133 @@ class TrajectoryPlotter(object):
         # reset buffer
         del self.frames
         self.frames = None
+
+        # close plotter
+        plt.close()
+
+        # clean
+        del self
+
+class LearningCurvePlotter(object):
+    """Plotter for learning curve.
+
+    Initializes ``fig``, ``ax`` and ``data``.
+
+    Parameters
+    ----------
+    axes_args: list
+        Lists of axis properties. The first element of each entry is the ``x_label``, the second is ``y_label``, the third is ``[y_limit_min, y_limit_max]`` and the fourth is ``y_scale``.
+    average_over: int, default=100
+        Number of points to average over.
+    percentiles: list, default=[25, 50, 75]
+        Percentile values for intraquartile ranges.
+    """
+
+    def __init__(self,
+        axis_args:list,
+        average_over:int=100,
+        percentiles:list=[25, 50, 75]
+    ):
+        """Class constructor for IQRPlotter."""
+
+        # set attributes
+        self.axis_args = axis_args
+        self.average_over = average_over
+        self.percentiles = percentiles
+
+        # turn on interactive mode
+        plt.ion()
+
+        # initialize plot
+        self.fig = plt.figure(figsize=(6.0, 3.0))
+        self.ax = plt.gca()
+        self.ax.set_xlabel(self.axis_args[0])
+        self.ax.set_ylabel(self.axis_args[1])
+        self.ax.set_ylim(ymin=self.axis_args[2][0], ymax=self.axis_args[2][1])
+        self.ax.set_yscale(self.axis_args[3])
+        self.fig.tight_layout()
+
+        # initialze buffer
+        self.data = list()
+        self.line = None
+        self.line_faint = None
+
+    def add_data(self,
+        data_rewards:np.ndarray,
+        renew:bool=False,
+        color:str='k',
+        style:str='-'
+    ):
+        """Method to add reward data.
+
+        Parameters
+        ----------
+        data_rewards: :class:`numpy.ndarray`
+            Reward data for the plot.
+        renew: bool, default=False
+            Option to create a new list
+        color: str, default='k'
+            Line color of the plot.
+        style: str, default='-'
+            Line style of the plot.
+        """
+
+        # if averaging opted
+        if self.average_over is not None:
+            data_rewards_smooth = np.convolve(data_rewards, np.ones((self.average_over, )) / float(self.average_over), mode='same')
+
+        # update data
+        if renew:
+            self.data = list()
+            self.line = None
+            self.line_faint = None
+        self.data.append(data_rewards_smooth)
+        if self.line_faint is not None:
+            self.line_faint.remove()
+        if self.line is not None:
+            self.line.remove()
+        xs = list(range(len(self.data[0])))
+
+        # if single entry
+        if len(self.data) == 1:
+            q_mean = data_rewards_smooth
+            self.line_faint = self.ax.plot(xs, data_rewards, c=color, alpha=0.1, linewidth=0.5)[0]
+        else:
+            # if interquartile ranges given
+            if self.percentiles is not None:
+                q_min, q_mean, q_max = np.percentile(self.data, self.percentiles, 0)
+                self.line_faint = self.ax.fill_between(xs, q_min, q_max, facecolor=color, alpha=0.1)
+            # calculate mean
+            else:
+                q_mean = np.mean(self.data, 0)
+                self.line_faint = self.ax.plot(xs, q_mean, c=color, alpha=0.1, linewidth=0.5)[0]
+
+        # add new lines
+        self.line = self.ax.plot(xs, q_mean, c=color, linestyle=style, linewidth=1.5)[0]
+
+    def save_plot(self,
+        file_name:str
+    ):
+        """Method to save the plot.
+
+        Parameters
+        ----------
+        file_name: str
+            Name of the file without its extension.
+        """
+
+        plt.savefig(file_name)
+
+    def hold_plot(self):
+        """Method to hold the plot."""
+
+        # turn off interactive mode
+        plt.ioff()
+        # show plot
+        plt.show()
+
+    def close(self):
+        """Method to close the plotter."""
 
         # close plotter
         plt.close()